pdev:set

+++ index.js
@@ -3,6 +3,8 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var util = require('util');
 var ldap = require('ldapjs');
 var poolId = 0;
+var log = console.log.bind(console, ' => [ldap-pool] =>');
+var logError = console.error.bind(console, ' => [ldap-pool] => warning =>');
 function createTimeout(pool, client, timeout) {
     client.__inactiveTimeoutX = setTimeout(function () {
         var isDestroyable = false;
@@ -30,6 +32,11 @@ function clearInactive(pool, c) {
         return v !== c;
     });
 }
+function logSize(pool) {
+    log('active clients count => ', pool.active.length);
+    log('inactive clients count => ', pool.inactive.length);
+    log('total clients count => ', pool.inactive.length + pool.active.length);
+}
 var Pool = (function () {
     function Pool(opts) {
         this.id = ++poolId;
@@ -53,35 +60,46 @@ var Pool = (function () {
         var client = ldap.createClient(this.connOpts);
         client.cdtClientId = this.clientId++;
         client.on('idle', function () {
-            console.log("client with id => " + client.cdtClientId + " is idle.");
+            if (client.ldapPoolRemoved) {
+                return;
+            }
+            log("client with id => " + client.cdtClientId + " is idle.");
+            logSize(_this);
             client.ldapPoolRemoved = true;
             clearActive(_this, client);
             clearInactive(_this, client);
             _this.addClient();
             client.unbind(function () {
                 client.destroy();
+                client.removeAllListeners();
             });
         });
         client.on('error', function (e) {
-            console.error(" => LDAP client error (in client pool, id=" + client.cdtClientId + ") => ", e.stack || e);
+            if (client.ldapPoolRemoved) {
+                return;
+            }
+            logError("client error (in client pool, id=" + client.cdtClientId + ") => \n", e.stack || e);
+            logSize(_this);
             client.ldapPoolRemoved = true;
             clearActive(_this, client);
             clearInactive(_this, client);
             _this.addClient();
             client.unbind(function () {
                 client.destroy();
+                client.removeAllListeners();
             });
         });
         client.bind(this.dn, this.pwd, function (err) {
             if (err) {
-                console.error(err);
+                logError('\n', err.stack || err);
             }
             else {
-                console.log('Successfully bound client.');
+                log('Successfully bound client.');
             }
         });
         this.inactive.push(client);
         client.returnToPool = function () {
+            logSize(_this);
             if (client.ldapPoolRemoved) {
                 return;
             }
@@ -97,6 +115,7 @@ var Pool = (function () {
     };
     Pool.prototype.getClient = function () {
         var _this = this;
+        logSize(this);
         var c = this.inactive.pop();
         if (c) {
             this.active.unshift(c);
@@ -109,6 +128,7 @@ var Pool = (function () {
         }
     };
     Pool.prototype.getClientSync = function () {
+        logSize(this);
         var c;
         if (c = this.inactive.pop()) {
             clearTimeout(c.__inactiveTimeoutX);
@@ -119,6 +139,7 @@ var Pool = (function () {
         return this.active[oldestActive];
     };
     Pool.prototype.returnClientToPool = function (c) {
+        logSize(this);
         if (c.ldapPoolRemoved) {
             return;
         }
